refactor: badge 컴포넌트 로직 분리 (#ATR-272)
