+++ packages/design-system/src/Badge/Badge.type.ts
@@ -1,0 +1,4 @@
+import { type VariantProps } from 'class-variance-authority'
+import badgeVariants from './Badge.style'
+
+export type BadgeVaraintsProps = VariantProps<typeof badgeVariants>
